refactor: cria resource para padronizar retorno de erros
