+++ app/Infrastructure/Exception/Enum/ErrorCodeEnum.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Infrastructure\Exception;
+namespace App\Infrastructure\Exception\Enum;
 
 enum ErrorCodeEnum: string
 {
