+++ app/Infrastructure/Exception/Handler/DomainExceptionHandler.php
@@ -4,7 +4,8 @@ declare(strict_types=1);
 
 namespace App\Infrastructure\Exception\Handler;
 
-use App\Infrastructure\Exception\ErrorCodeEnum;
+use App\Infrastructure\Exception\Enum\ErrorCodeEnum;
+use App\Infrastructure\Exception\Resource\ErrorResource;
 use DomainException;
 use Hyperf\Contract\StdoutLoggerInterface;
 use Hyperf\ExceptionHandler\ExceptionHandler;
@@ -15,8 +16,9 @@ use Throwable;
 
 class DomainExceptionHandler extends ExceptionHandler
 {
-    public function __construct(protected StdoutLoggerInterface $logger)
-    {
+    public function __construct(
+        private StdoutLoggerInterface $logger,
+    ) {
     }
 
     public function handle(Throwable $throwable, ResponseInterface $response)
@@ -24,12 +26,14 @@ class DomainExceptionHandler extends ExceptionHandler
         $this->logger->error(sprintf('%s[%s] in %s', $throwable->getMessage(), $throwable->getLine(), $throwable->getFile()));
         $this->logger->error($throwable->getTraceAsString());
 
+        $resource = new ErrorResource(
+            errorCode: ErrorCodeEnum::DOMAIN,
+            throwable: $throwable,
+        );
         return $response
-            ->withStatus(Status::BAD_REQUEST)->withBody(new SwooleStream(json_encode([
-                'error' => [
-                    'code' => ErrorCodeEnum::DOMAIN->value,
-                    'message' => $throwable->getMessage(),
-                ],
+            ->withStatus(Status::BAD_REQUEST)
+            ->withBody(new SwooleStream(json_encode([
+                'error' => $resource->toArray()
             ])));
     }
